Revision: 200951 Kit: 201001

+++ messagingappbase/mce/src/MceDeliveryReportView.cpp
@@ -180,7 +180,10 @@ void CMceDeliveryReportView::DynInitMenuPaneL( TInt aResourceId, CEikMenuPane* a
             if(  ! MceAppUi()->LogEngine()->
                     IsNumberAvailable( iContainer->ListBox()->CurrentItemIndex() ) )
                 {
-                aMenuPane->DeleteMenuItem( EMceCmdCall );
+					if ( aMenuPane->MenuItemExists( EMceCmdCall, aPosition ) ) 
+					{
+						aMenuPane->DeleteMenuItem( EMceCmdCall );
+					}
                 }
 
             if ( iContainer->ListBox()->Model()->NumberOfItems() == 0 )
